requirements.txt
nltk
streamlit
pandas
matplotlib
seaborn
scikit-learn
py2neo
neo4j~=5.28.0
pyvis
tqdm
python-dotenv
numpy<2.0
spacy==3.7.2
thinc==8.2.2
blis==0.7.11
